Update README.md
Due the penetration of the internet in all domains of life which has led to increase of people’s participation actively
and give remarks as an issue of communicating their concern/feedback/opinion in various online forums. Although
most of the times these comments are helpful for the creator toextemporize the substance that is being provided to
people, but sometimes these may be abusive and create hatred-feeling among the people. Thus as these are openly
available to the public which is being viewed from various sections of the society, people in different age groups,
different communities and different socio-economic background, it becomes the prime responsibility of the
content-creator ( the host) to filter out these comments in order to stop the spread of negativity or hatred within
people.
Lately there has been many cases in which the growing menace of hate and negativity has been witnessed in the
online platforms especially social media as such, many governments around the world has seen the rise of cases
related to cyber bullying that has led to spread of hatred and violence.
Since the democratization of substance creation following the dispatch of web-based media stages, every single
one of us has become content makers making and distributing our own substance, which thus has made a
framework where the nature of distributed substance cannot, at this point be controlled. The effect of the most
recent twenty years' innovation unrest is presently affecting organizations, political frameworks, family lives,
society, and individuals 
.
Detecting Toxic comments has been a great challenge for the all the scholars in the field of research and
development. This domain has drawn lot of interests not just because of the spread of hate but also people refraining
people from participating in online forums which diversely affects for all the creators/content-providers to provide
a relief to engage in a healthy public interaction which can be accessed by public without any hesitation.
There have been sure turns of developments in this area which includes couple of models served through API. But
the models still make errors and still fail to provide an accurate solution to the problem and may need more clarification and develpment in future